requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
certifi==2024.7.4
    # via requests
charset-normalizer==3.3.2
    # via requests
et-xmlfile==1.1.0
    # via openpyxl
idna==3.7
    # via requests
numpy==2.0.0
    # via pandas
openpyxl==3.1.5
    # via -r requirements.in
pandas==2.2.2
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.0.1
    # via -r requirements.in
pytz==2024.1
    # via pandas
requests==2.32.3
    # via -r requirements.in
six==1.16.0
    # via python-dateutil
tzdata==2024.1
    # via pandas
urllib3==2.2.2
    # via requests
